인가처리

+++ back/routes/post.js
@@ -157,6 +157,9 @@ router.delete('/:id', isLoggedIn, async (req, res, next) => {
         if (!post) {
             return res.status(404).send('포스트가 존재하지 않습니다.');
         }
+        if(req.user.id!==req.params.postId){
+            return res.status(403).send('삭제 할 권한이 없습니다.');
+        }
         await db.Post.update( {isDeleted:true},{where: { id: req.params.id } });
         res.send(req.params.id);
     } catch (e) {
@@ -218,6 +221,9 @@ router.post('/:id/comment', isLoggedIn, async (req, res, next) => { // POST /api
 
 router.patch('/:id/comment', isLoggedIn, async (req, res, next) => {
     try {
+        if(req.user.id!==req.params.postId){
+            return res.status(403).send('수정 할 권한이 없습니다.');
+        }
         await db.Comment.update({ content: req.body.content
         }, { where: {
                 id: req.params.id,
@@ -233,6 +239,9 @@ router.patch('/:id/comment', isLoggedIn, async (req, res, next) => {
 
 router.delete('/:id/comment', isLoggedIn, async (req, res, next) => {
     try {
+        if(req.user.id!==req.params.postId){
+            return res.status(403).send('삭제 할 권한이 없습니다.');
+        }
         await db.Comment.destroy({ where: { id: req.params.id, UserId: req.user.id, } });
         res.status(200).json(req.params.id);
     } catch (e) {
@@ -243,7 +252,6 @@ router.delete('/:id/comment', isLoggedIn, async (req, res, next) => {
 
 router.post('/:id/like', isLoggedIn, async (req, res, next) => {
     try {
-        console.log('조아요');
         const post = await db.Post.findOne({ where: { id: req.params.id }});
         if (!post) {
             return res.status(404).send('포스트가 존재하지 않습니다.');
@@ -258,6 +266,9 @@ router.post('/:id/like', isLoggedIn, async (req, res, next) => {
 
 router.delete('/:id/like', isLoggedIn, async (req, res, next) => {
     try {
+        if(req.user.id!==req.params.postId){
+            return res.status(403).send('수정 할 권한이 없습니다.');
+        }
         const post = await db.Post.findOne({ where: { id: req.params.id }});
         if (!post) {
             return res.status(404).send('포스트가 존재하지 않습니다.');
